Ignore 05_TearDown
Tear Down assumes that there are two resource groups, but does not apply the --yes flag to the calls causing it to hang. Given that all it does is remove resource groups (without verifying if they are there) and removes the docker image this notebook can be ignored because the default cleanup task will do these same very tasks.